// gran-turismo-positions/src/types/Driver.ts
export interface Driver {
  carId?: number;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  position?: number;
  classPosition?: number;
  isCarOnTrack?: boolean;
  iRating?: number;
  iRatingChange?: number;
  carClassShortName?: string;
  iRacingLicString?: string;
  iRacingLicSubLevel?: number;
  spectateCarId?: number;
}